Experience and Tiers
I changed how the item levels up. Added a tier element which helps determine how hard it is to level up. Some factors will need to be defined outside of the constructor. I settled on this because I want the Arm/Legs/Implant Type to influence the baseStats/BestStats/etc